add a wrapper msg-emulator for generic self-talking or thinking
